Refactored data processing
