+++ neuro_ai/neuro_ai.py
@@ -3,6 +3,9 @@ import socketio
 from loguru import logger
 from ccl.behaviors.concentration import Concentration
 from ccl.models.eeg import InputData
+from ccl.pre_processing.data_processing import DataProcessig
+from ccl.pre_processing.pre_procesing import setup_mne_data
+import sys
 
 # Create a FastAPI app
 app = FastAPI()
@@ -18,6 +21,9 @@ namespace = "/eeg"
 
 # Create the behaviors objects
 concentration = Concentration()
+
+#TODO: Create data object
+data_processor = DataProcessig()
 
 
 @sio.event(namespace=namespace)
@@ -67,8 +73,18 @@ async def eegData(sid, data: dict):
         # logger.error("Invalid data type")
         return
 
-    processed_data = concentration.pre_process_data(input_data.data.data)  # DATAAAAAAAA
-    concentration_level = concentration.concentration_level(processed_data)
+
+    #TODO: Trasform to nme data
+    mne_raw = setup_mne_data(input_data.data.data)
+
+    #TODO: Calculate and keep psd
+    data_processor.process_psd_data(mne_raw)
+
+    # concentration
+    #TODO: adjust to take new data format waves
+    #processed_data = concentration.pre_process_data(mne_raw)  # DATAAAAAAAA
+    concentration_level = concentration.concentration_level(data_processor.wave_data["psd_power"]["alpha"][-1:],
+                                                            data_processor.wave_data["psd_power"]["beta"][-1:])
 
     logger.info(f"Concentration Level: {concentration_level.value}")
 
@@ -81,6 +97,14 @@ async def eegData(sid, data: dict):
 
 
 if __name__ == "__main__":
-    import uvicorn
+    #TODO: Setup api or not
+    run_api = True
+    if len(sys.argv) > 1:
+        run_api = sys.argv[1]
 
-    uvicorn.run(app, host="0.0.0.0", port=9090)
+    if run_api:
+        import uvicorn
+        uvicorn.run(app, host="0.0.0.0", port=9090)
+    else:
+        #TODO: Adjust to be able to run without the API
+        print("Comming soon")
